+++ brion/synapse.cpp
@@ -21,7 +21,7 @@
  */
 
 #include "synapse.h"
-#include "detail/lockHDF5.h"
+#include "detail/hdf5Mutex.h"
 
 #include <highfive/H5DataSet.hpp>
 #include <highfive/H5File.hpp>
@@ -32,7 +32,6 @@
 #include <boost/regex.hpp>
 
 #include <lunchbox/log.h>
-#include <lunchbox/scopedMutex.h>
 
 #include <bitset>
 #include <fstream>
@@ -85,7 +84,7 @@ class SynapseFile : public boost::noncopyable
 public:
     explicit SynapseFile(const std::string& source)
     {
-        lunchbox::ScopedWrite mutex(detail::hdf5Lock());
+        std::lock_guard<std::mutex> lock(detail::hdf5Mutex());
 
         try
         {
@@ -116,7 +115,7 @@ public:
 
     ~SynapseFile()
     {
-        lunchbox::ScopedWrite mutex(detail::hdf5Lock());
+        std::lock_guard<std::mutex> lock(detail::hdf5Mutex());
         _file.reset();
     }
 
@@ -127,7 +126,7 @@ public:
         if (!bits.any())
             return SynapseMatrix();
 
-        lunchbox::ScopedWrite mutex(detail::hdf5Lock());
+        std::lock_guard<std::mutex> lock(detail::hdf5Mutex());
         Dataset dataset;
         if (!_openDataset(gid, dataset))
             return SynapseMatrix();
@@ -149,7 +148,7 @@ public:
     size_t getNumAttributes() const { return _numAttributes; }
     size_t getNumSynapses(const GIDSet& gids) const
     {
-        lunchbox::ScopedWrite mutex(detail::hdf5Lock());
+        std::lock_guard<std::mutex> lock(detail::hdf5Mutex());
         size_t numSynapses = 0;
         for (const uint32_t gid : gids)
         {
@@ -376,7 +375,7 @@ private:
         // usually results in waiting for I/O and non-parallizable search thanks
         // to HDF5
 
-        lunchbox::ScopedWrite mutex(detail::hdf5Lock());
+        std::lock_guard<std::mutex> lock(detail::hdf5Mutex());
         HighFive::SilenceHDF5 silence;
 
         // this trial-and-error is the 'fastest' path found